Move blockentity mixins to a separate package
