+++ patchwork-extensions-block/src/main/java/net/patchworkmc/mixin/extensions/block/MixinBlock.java
@@ -27,9 +27,6 @@ import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import net.minecraftforge.common.extensions.IForgeBlock;
 
 import net.minecraft.block.Block;
@@ -41,8 +38,6 @@ import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.CollisionView;
 
-import net.patchworkmc.impl.extensions.block.BlockContext;
-
 @Mixin(Block.class)
 public class MixinBlock implements IForgeBlock {
 	@Shadow
@@ -53,11 +48,6 @@ public class MixinBlock implements IForgeBlock {
 	private Set<Identifier> cachedTags;
 	@Unique
 	private int tagVersion;
-
-	@Inject(method = "hasBlockEntity", at = @At("RETURN"), cancellable = true)
-	public void patchwork_hasBlockEntity(CallbackInfoReturnable<Boolean> info) {
-		info.setReturnValue(BlockContext.block_hasBlockEntity(this));
-	}
 
 	@Override
 	public float getSlipperiness(BlockState state, CollisionView world, BlockPos pos, Entity entity) {
