+++ patchwork-extensions-block/src/main/java/net/patchworkmc/mixin/extensions/block/blockentity/MixinAbstractRedstoneGateBlock.java
@@ -17,33 +17,24 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.patchworkmc.mixin.extensions.block;
+package net.patchworkmc.mixin.extensions.block.blockentity;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
+import net.minecraft.block.AbstractRedstoneGateBlock;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.client.render.chunk.ChunkRenderer;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
 
 import net.patchworkmc.impl.extensions.block.BlockContext;
-import net.patchworkmc.impl.extensions.block.Signatures;
 
-@Mixin(ChunkRenderer.class)
-public abstract class MixinChunkRenderer {
-	private static final ThreadLocal<Object> rebuildChunk_blockState = BlockContext.createContext();
-	// Block block = blockState.getBlock();
-	@Redirect(method = "rebuildChunk", at = @At(value = "INVOKE", target = Signatures.BlockState_getBlock, ordinal = 0))
-	public Block patchwork_rebuildChunk_getBlock(BlockState blockstate) {
-		BlockContext.setContext(rebuildChunk_blockState, blockstate);
-		return blockstate.getBlock();
-	}
-
-	// if (block.hasBlockEntity()) {
-	@Redirect(method = "rebuildChunk", at = @At(value = "INVOKE", target = Signatures.Block_hasBlockEntity, ordinal = 0))
-	public boolean patchwork_rebuildChunk_hasBlockEntity(Block dummy) {
-		BlockState blockState = BlockContext.releaseContext(rebuildChunk_blockState);
-		return BlockContext.hasBlockEntity(blockState);
+@Mixin(AbstractRedstoneGateBlock.class)
+public abstract class MixinAbstractRedstoneGateBlock {
+	@Redirect(method = "neighborUpdate", at = @At(value = "INVOKE", target = "net/minecraft/block/AbstractRedstoneGateBlock.hasBlockEntity()Z", ordinal = 0))
+	private boolean patchwork_neighborUpdate_hasBlockEntity(AbstractRedstoneGateBlock dummy, BlockState state, World world, BlockPos pos, Block block, BlockPos neighborPos, boolean moved) {
+		return BlockContext.hasBlockEntity(state);
 	}
 }
