+++ patchwork-extensions-block/src/main/java/net/patchworkmc/mixin/extensions/block/blockentity/MixinChunkHolder.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.patchworkmc.mixin.extensions.block;
+package net.patchworkmc.mixin.extensions.block.blockentity;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -25,15 +25,17 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.SpongeBlock;
+import net.minecraft.server.world.ChunkHolder;
 
 import net.patchworkmc.impl.extensions.block.BlockContext;
 import net.patchworkmc.impl.extensions.block.Signatures;
 
-@Mixin(SpongeBlock.class)
-public abstract class MixinSpongeBlock {
-	@Redirect(method = "absorbWater", at = @At(value = "INVOKE", target = Signatures.BlockState_getBlock, ordinal = 3))
-	private Block patchwork_absorbWater_getBlock(BlockState blockstate) {
-		return BlockContext.hasBlockEntityBlockMarker(blockstate);
+@Mixin(ChunkHolder.class)
+public class MixinChunkHolder {
+	// if (world.getBlockState(blockPos).getBlock().hasBlockEntity()) {
+	// if (world.getBlockState(blockPos2).getBlock().hasBlockEntity()) {
+	@Redirect(method = "flushUpdates", at = @At(value = "INVOKE", target = Signatures.BlockState_getBlock))
+	private Block patchwork_flushUpdates_getBlock(BlockState blockState) {
+		return BlockContext.hasBlockEntityBlockMarker(blockState);
 	}
 }
