+++ patchwork-extensions-block/src/main/java/net/patchworkmc/mixin/extensions/block/blockentity/MixinFallingBlockEntity.java
@@ -17,25 +17,27 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.patchworkmc.mixin.extensions.block;
+package net.patchworkmc.mixin.extensions.block.blockentity;
 
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.Constant;
+import org.spongepowered.asm.mixin.injection.ModifyConstant;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.server.world.ChunkHolder;
+import net.minecraft.block.BlockEntityProvider;
+import net.minecraft.entity.FallingBlockEntity;
 
 import net.patchworkmc.impl.extensions.block.BlockContext;
-import net.patchworkmc.impl.extensions.block.Signatures;
 
-@Mixin(ChunkHolder.class)
-public class MixinChunkHolder {
-	// if (world.getBlockState(blockPos).getBlock().hasBlockEntity()) {
-	// if (world.getBlockState(blockPos2).getBlock().hasBlockEntity()) {
-	@Redirect(method = "flushUpdates", at = @At(value = "INVOKE", target = Signatures.BlockState_getBlock))
-	public Block patchwork_flushUpdates_getBlock(BlockState blockState) {
-		return BlockContext.hasBlockEntityBlockMarker(blockState);
+@Mixin(FallingBlockEntity.class)
+public abstract class MixinFallingBlockEntity {
+	////////////////////////
+	/// tick()
+	////////////////////////
+	// } else if (block2 != block && block2 instanceof BlockEntityProvider) {
+	@ModifyConstant(method = "tick", constant = @Constant(classValue = BlockEntityProvider.class, ordinal = 0))
+	private boolean patchwork_tick_instanceof_BlockEntityProvider(Object object, Class<?> clazz) {
+		FallingBlockEntity me = (FallingBlockEntity) (Object) this;
+		// Forge: return ((Block) object).hasBlockEntity();
+		return BlockContext.hasBlockEntity(me.getBlockState());
 	}
 }
