+++ patchwork-extensions-block/src/main/java/net/patchworkmc/mixin/extensions/block/blockentity/MixinHopperBlockEntity.java
@@ -17,37 +17,28 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.patchworkmc.mixin.extensions.block;
-
-import java.util.Iterator;
+package net.patchworkmc.mixin.extensions.block.blockentity;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.At.Shift;
-import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.entity.HopperBlockEntity;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.explosion.Explosion;
+import net.minecraft.world.World;
 
 import net.patchworkmc.impl.extensions.block.BlockContext;
 import net.patchworkmc.impl.extensions.block.Signatures;
 
-@Mixin(Explosion.class)
-public abstract class MixinExplosion {
-	private static final ThreadLocal<Object> affectWorld_blockState = BlockContext.createContext();
-	@Inject(method = "affectWorld", locals = LocalCapture.CAPTURE_FAILHARD, at = @At(value = "INVOKE", shift = Shift.BEFORE, target = Signatures.Block_hasBlockEntity, ordinal = 0))
-	private void patchwork_affectWorld_hasBlockEntity_before(boolean bl, CallbackInfo ci, boolean bl2, Iterator var3, BlockPos blockPos, BlockState blockState, Block block) {
-		BlockContext.setContext(affectWorld_blockState, blockState);
-	}
-
-	@Redirect(method = "affectWorld", at = @At(value = "INVOKE", target = Signatures.Block_hasBlockEntity, ordinal = 0))
-	public boolean patchwork_affectWorld_hasBlockEntity(Block dummy) {
-		BlockState blockState = BlockContext.releaseContext(affectWorld_blockState);
+@Mixin(HopperBlockEntity.class)
+public abstract class MixinHopperBlockEntity {
+	// } else if (block.hasBlockEntity()) {
+	@Redirect(method = "getInventoryAt(Lnet/minecraft/world/World;DDD)Lnet/minecraft/inventory/Inventory;", at = @At(value = "INVOKE", target = Signatures.Block_hasBlockEntity, ordinal = 0))
+	private static boolean patchwork_getInventoryAt_hasBlockEntity(Block dummy, World world, double x, double y, double z) {
+		BlockPos blockPos = new BlockPos(x, y, z);
+		BlockState blockState = world.getBlockState(blockPos);
 		return BlockContext.hasBlockEntity(blockState);
 	}
 }
