+++ patchwork-extensions-block/src/main/java/net/patchworkmc/mixin/extensions/block/blockentity/MixinMinecraftClient.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.patchworkmc.mixin.extensions.block;
+package net.patchworkmc.mixin.extensions.block.blockentity;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -25,20 +25,21 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.entity.HopperBlockEntity;
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
 
 import net.patchworkmc.impl.extensions.block.BlockContext;
 import net.patchworkmc.impl.extensions.block.Signatures;
 
-@Mixin(HopperBlockEntity.class)
-public abstract class MixinHopperBlockEntity {
-	// } else if (block.hasBlockEntity()) {
-	@Redirect(method = "getInventoryAt(Lnet/minecraft/world/World;DDD)Lnet/minecraft/inventory/Inventory;", at = @At(value = "INVOKE", target = Signatures.Block_hasBlockEntity, ordinal = 0))
-	private static boolean patchwork_getInventoryAt_hasBlockEntity(Block dummy, World world, double x, double y, double z) {
-		BlockPos blockPos = new BlockPos(x, y, z);
-		BlockState blockState = world.getBlockState(blockPos);
+@Mixin(MinecraftClient.class)
+public abstract class MixinMinecraftClient {
+	// if (bl && Screen.hasControlDown() && block.hasBlockEntity()) {
+	@Redirect(method = "doItemPick", at = @At(value = "INVOKE", target = Signatures.Block_hasBlockEntity, ordinal = 0))
+	private boolean patchwork_doItemPick_hasBlockEntity(Block dummy) {
+		MinecraftClient me = (MinecraftClient) (Object) this;
+		BlockPos blockPos = ((BlockHitResult) me.crosshairTarget).getBlockPos();
+		BlockState blockState = me.world.getBlockState(blockPos);
 		return BlockContext.hasBlockEntity(blockState);
 	}
 }
