+++ patchwork-extensions-block/src/main/java/net/patchworkmc/mixin/extensions/block/blockentity/MixinWorldChunk.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.patchworkmc.mixin.extensions.block;
+package net.patchworkmc.mixin.extensions.block.blockentity;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
@@ -66,7 +66,7 @@ public abstract class MixinWorldChunk {
 	/// createBlockEntity()
 	////////////////////////
 	@Inject(method = "createBlockEntity", locals = LocalCapture.CAPTURE_FAILHARD, cancellable = true, at = @At(value = "INVOKE", target = Signatures.BlockState_getBlock))
-	public void patchwork_createBlockEntity(BlockPos blockPos, CallbackInfoReturnable<BlockEntity> info, BlockState blockState) {
+	private void patchwork_createBlockEntity(BlockPos blockPos, CallbackInfoReturnable<BlockEntity> info, BlockState blockState) {
 		info.setReturnValue(patchwork_createBlockEntity(blockState));
 		info.cancel();
 	}
@@ -79,7 +79,7 @@ public abstract class MixinWorldChunk {
 	//     blockEntity3 = ((BlockEntityProvider)block).createBlockEntity(this.world);
 	private static final ThreadLocal<Object> loadBlockEntity_blockEntity = BlockContext.createContext();
 	@Redirect(method = "loadBlockEntity", at = @At(value = "INVOKE", target = Signatures.BlockState_getBlock, ordinal = 0))
-	public Block patchwork_loadBlockEntity_getBlock(BlockState blockState) {
+	private Block patchwork_loadBlockEntity_getBlock(BlockState blockState) {
 		BlockEntity blockEntity = patchwork_createBlockEntity(blockState);
 
 		if (blockEntity != null) {
@@ -90,7 +90,7 @@ public abstract class MixinWorldChunk {
 	}
 
 	@Redirect(method = "loadBlockEntity", at = @At(value = "INVOKE", target = Signatures.BlockEntityProvider_createBlockEntity, ordinal = 0))
-	public BlockEntity patchwork_loadBlockEntity_createBlockEntity(BlockEntityProvider dummy, BlockView view) {
+	private BlockEntity patchwork_loadBlockEntity_createBlockEntity(BlockEntityProvider dummy, BlockView view) {
 		BlockEntity blockEntity = BlockContext.releaseContext(loadBlockEntity_blockEntity);
 		return blockEntity;
 	}
@@ -101,46 +101,46 @@ public abstract class MixinWorldChunk {
 	private static final ThreadLocal<Object> loadBlockEntity_blockState2 = BlockContext.createContext();
 	// } else if (block2 != block && block2 instanceof BlockEntityProvider) {
 	@Inject(method = "setBlockState", locals = LocalCapture.CAPTURE_FAILHARD, at = @At(value = "CONSTANT", args = Signatures.PATCHWORK_YARN_CLS_BLOCKENTITYPROVIDER, ordinal = 0), require = 0)
-	public void patchwork_yarn_setBlockState_instanceof_BlockEntityProvider(BlockPos pos, BlockState state, boolean bl, CallbackInfoReturnable<BlockState> cir, int i, int j, int k, ChunkSection chunkSection, boolean bl2, BlockState blockState, Block block, Block block2) {
+	private void patchwork_yarn_setBlockState_instanceof_BlockEntityProvider(BlockPos pos, BlockState state, boolean bl, CallbackInfoReturnable<BlockState> cir, int i, int j, int k, ChunkSection chunkSection, boolean bl2, BlockState blockState, Block block, Block block2) {
 		BlockContext.setContext(loadBlockEntity_blockState2, blockState);
 	}
 
 	@Inject(method = "setBlockState", locals = LocalCapture.CAPTURE_FAILHARD, at = @At(value = "CONSTANT", args = Signatures.PATCHWORK_REOBF_CLS_BLOCKENTITYPROVIDER, ordinal = 0), require = 0)
-	public void patchwork_reobf_setBlockState_instanceof_BlockEntityProvider(BlockPos pos, BlockState state, boolean bl, CallbackInfoReturnable<BlockState> cir, int i, int j, int k, ChunkSection chunkSection, boolean bl2, BlockState blockState, Block block, Block block2) {
+	private void patchwork_reobf_setBlockState_instanceof_BlockEntityProvider(BlockPos pos, BlockState state, boolean bl, CallbackInfoReturnable<BlockState> cir, int i, int j, int k, ChunkSection chunkSection, boolean bl2, BlockState blockState, Block block, Block block2) {
 		BlockContext.setContext(loadBlockEntity_blockState2, blockState);
 	}
 
 	@ModifyConstant(method = "setBlockState", constant = @Constant(classValue = BlockEntityProvider.class, ordinal = 0))
-	public boolean patchwork_setBlockState_instanceof_BlockEntityProvider(Object object, Class<?> clazz) {
+	private boolean patchwork_setBlockState_instanceof_BlockEntityProvider(Object object, Class<?> clazz) {
 		BlockState blockState2 = BlockContext.releaseContext(loadBlockEntity_blockState2);
 		return BlockContext.hasBlockEntity(blockState2);
 	}
 
 	// if (block2 instanceof BlockEntityProvider) {
 	@Inject(method = "setBlockState", locals = LocalCapture.CAPTURE_FAILHARD, at = @At(value = "CONSTANT", args = Signatures.PATCHWORK_YARN_CLS_BLOCKENTITYPROVIDER, ordinal = 1), require = 0)
-	public void patchwork_yarn_setBlockState_instanceof_BlockEntityProvider_1(BlockPos pos, BlockState state, boolean bl, CallbackInfoReturnable<BlockState> cir, int i, int j, int k, ChunkSection chunkSection, boolean bl2, BlockState blockState, Block block, Block block2) {
+	private void patchwork_yarn_setBlockState_instanceof_BlockEntityProvider_1(BlockPos pos, BlockState state, boolean bl, CallbackInfoReturnable<BlockState> cir, int i, int j, int k, ChunkSection chunkSection, boolean bl2, BlockState blockState, Block block, Block block2) {
 		BlockContext.setContext(loadBlockEntity_blockState2, blockState);
 	}
 
 	@Inject(method = "setBlockState", locals = LocalCapture.CAPTURE_FAILHARD, at = @At(value = "CONSTANT", args = Signatures.PATCHWORK_REOBF_CLS_BLOCKENTITYPROVIDER, ordinal = 1), require = 0)
-	public void patchwork_reobf_setBlockState_instanceof_BlockEntityProvider_1(BlockPos pos, BlockState state, boolean bl, CallbackInfoReturnable<BlockState> cir, int i, int j, int k, ChunkSection chunkSection, boolean bl2, BlockState blockState, Block block, Block block2) {
+	private void patchwork_reobf_setBlockState_instanceof_BlockEntityProvider_1(BlockPos pos, BlockState state, boolean bl, CallbackInfoReturnable<BlockState> cir, int i, int j, int k, ChunkSection chunkSection, boolean bl2, BlockState blockState, Block block, Block block2) {
 		BlockContext.setContext(loadBlockEntity_blockState2, blockState);
 	}
 
 	@ModifyConstant(method = "setBlockState", constant = @Constant(classValue = BlockEntityProvider.class, ordinal = 1))
-	public boolean patchwork_setBlockState_instanceof_BlockEntityProvider_1(Object object, Class<?> clazz) {
+	private boolean patchwork_setBlockState_instanceof_BlockEntityProvider_1(Object object, Class<?> clazz) {
 		BlockState blockState2 = BlockContext.releaseContext(loadBlockEntity_blockState2);
 		return BlockContext.hasBlockEntity(blockState2);
 	}
 
 	// if (block instanceof BlockEntityProvider) {
 	@ModifyConstant(method = "setBlockState", constant = @Constant(classValue = BlockEntityProvider.class, ordinal = 2))
-	public boolean patchwork_setBlockState_instanceof_BlockEntityProvider_2(Object object, Class<?> clazz, BlockPos pos, BlockState state, boolean bl) {
+	private boolean patchwork_setBlockState_instanceof_BlockEntityProvider_2(Object object, Class<?> clazz, BlockPos pos, BlockState state, boolean bl) {
 		return BlockContext.hasBlockEntity(state);
 	}
 
 	@Redirect(method = "setBlockState", at = @At(value = "INVOKE", target = Signatures.BlockEntityProvider_createBlockEntity, ordinal = 0))
-	public BlockEntity patchwork_setBlockState_createBlockEntity(BlockEntityProvider invoker, BlockView view, BlockPos pos, BlockState state, boolean bl) {
+	private BlockEntity patchwork_setBlockState_createBlockEntity(BlockEntityProvider invoker, BlockView view, BlockPos pos, BlockState state, boolean bl) {
 		IForgeBlockState forgeBlockState = (IForgeBlockState) state;
 		return forgeBlockState.createTileEntity(view);
 	}
@@ -150,7 +150,7 @@ public abstract class MixinWorldChunk {
 	////////////////////////
 	// if (this.getBlockState(pos).getBlock() instanceof BlockEntityProvider) {
 	@Redirect(method = "setBlockEntity", at = @At(value = "INVOKE", target = Signatures.BlockState_getBlock, ordinal = 0))
-	public Block patchwork_setBlockEntity_getBlock(BlockState blockState) {
+	private Block patchwork_setBlockEntity_getBlock(BlockState blockState) {
 		return BlockContext.hasBlockEntityBlockMarker(blockState);
 	}
 }
