+++ patchwork-extensions-block/src/main/resources/patchwork-extensions-block-blockentity.mixins.json
@@ -1,0 +1,28 @@
+{
+  "required": true,
+  "package": "net.patchworkmc.mixin.extensions.block.blockentity",
+  "compatibilityLevel": "JAVA_8",
+  "mixins": [
+    "MixinAbstractRedstoneGateBlock",
+    "MixinBlockArgumentParser",
+    "MixinBlock",
+    "MixinChestBlockEntity",
+    "MixinChunkHolder",
+    "MixinChunkRegion",
+    "MixinExplosion",
+    "MixinHopperBlockEntity",
+    "MixinFallingBlockEntity",
+    "MixinPistonBlock",
+    "MixinSpongeBlock",
+    "MixinWorld",
+    "MixinWorldChunk"
+  ],
+  "client": [
+    "MixinChunkRenderer",
+    "MixinWorldRenderer",
+    "MixinMinecraftClient"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
